// modules/mmorpg/src/model/generated/server/demo/protocol/messages.ts
export interface MapHostEndpoint {
  name: string;
  ip: string;
  port: number;
  protocol: string;
  audience: string;
}
// ...
export interface DynamicMapAssignmentSnapshot {
  requestId: string;
  mapConfigId: number;
  mapInstanceId: bigint;
  channelId?: number;
  maxPlayers?: number;
  privateRoster?: PrivateMapRoster;
}
// ...
export interface S2MM_RegisterMapHost extends IRequest {
  rpcId?: number;
  endpoint: MapHostEndpoint;
  generation: bigint;
  staticMapCount: number;
  dynamicMapCount: number;
  playerCount: number;
  assignments: readonly DynamicMapAssignmentSnapshot[];
  maxMaps?: number;
  maxPlayers?: number;
  mapConfigIds: readonly number[];
}
// ...
export interface S2MM_MapHostHeartbeat extends IRequest {
  rpcId?: number;
  mapHostName: string;
  generation: bigint;
  staticMapCount: number;
  dynamicMapCount: number;
  playerCount: number;
  maxMaps?: number;
  maxPlayers?: number;
  mapConfigIds: readonly number[];
}
// ...
export interface MM2S_DynamicMapDisposed extends IResponse {
  message?: string;
  error?: number;
  rpcId?: number;
  accepted: boolean;
}
// ...
export interface PrivateMapRoster {
  characterIds: readonly bigint[];
}