// modules/mmorpg/src/model/generated/server/demo/protocol/rpcs.ts
export const MapHostControlProtocol = {
  Register: defineRpc<S2MM_RegisterMapHost, MM2S_RegisterMapHost>({
    name: "MapHostControl.Register",
    requestCode: MsgCode.S2MM_RegisterMapHost,
    responseCode: MsgCode.MM2S_RegisterMapHost,
    requestCodec: S2MM_RegisterMapHostCodec,
    responseCodec: MM2S_RegisterMapHostCodec,
  }),
  Heartbeat: defineRpc<S2MM_MapHostHeartbeat, MM2S_MapHostHeartbeat>({
    name: "MapHostControl.Heartbeat",
    requestCode: MsgCode.S2MM_MapHostHeartbeat,
    responseCode: MsgCode.MM2S_MapHostHeartbeat,
    requestCodec: S2MM_MapHostHeartbeatCodec,
    responseCodec: MM2S_MapHostHeartbeatCodec,
  }),
  CreateAssigned: defineRpc<MM2M_CreateAssignedDynamicMap, M2MM_CreateAssignedDynamicMap>({
    name: "MapHostControl.CreateAssigned",
    requestCode: MsgCode.MM2M_CreateAssignedDynamicMap,
    responseCode: MsgCode.M2MM_CreateAssignedDynamicMap,
    requestCodec: MM2M_CreateAssignedDynamicMapCodec,
    responseCodec: M2MM_CreateAssignedDynamicMapCodec,
  }),
  DynamicMapDisposed: defineRpc<S2MM_DynamicMapDisposed, MM2S_DynamicMapDisposed>({
    name: "MapHostControl.DynamicMapDisposed",
    requestCode: MsgCode.S2MM_DynamicMapDisposed,
    responseCode: MsgCode.MM2S_DynamicMapDisposed,
    requestCodec: S2MM_DynamicMapDisposedCodec,
    responseCodec: MM2S_DynamicMapDisposedCodec,
  }),
};